Dodanie klasy abstrakcyjnej OverclockableComponent, po której dziedziczą Processor i Memory
